1.增加一条dupfromaltstack指令
2.调整一些命名和代码位置，因为要和java 转换工具共用部分代码。
