+++ src/AntShares.Compiler.MSIL/AntsModule.cs
@@ -2,7 +2,7 @@
 using System.Security.Cryptography;
 using System.Text;
 
-namespace AntShares.Compiler.MSIL
+namespace AntShares.Compiler
 {
     public class AntsModule
     {
@@ -85,10 +85,10 @@ namespace AntShares.Compiler.MSIL
     public class AntsMethod
     {
         public string name;
-        public List<ILParam> paramtypes = new List<ILParam>();
+        public List<Param> paramtypes = new List<Param>();
         public string returntype;
         //临时变量
-        public List<ILParam> body_Variables = new List<ILParam>();
+        public List<Param> body_Variables = new List<Param>();
 
         //临时记录在此，会合并到一起
         public SortedDictionary<int, AntsCode> body_Codes = new SortedDictionary<int, AntsCode>();
